DLUHC-264 move MapComponent sub components into new folder
