+++ dluhc-component-library/src/stories/MapLayer.stories.tsx
@@ -1,10 +1,10 @@
 import { FeatureCollection } from "geojson";
 import { Options as FillOptions } from "ol/style/Fill";
 import { Options as StrokeOptions } from "ol/style/Stroke";
-import BaseMap from "src/components/maps/BaseMap";
-import DrawingLayer from "src/components/maps/DrawingLayer";
-import MapContainer from "src/components/maps/MapContainer";
-import MapLayer from "src/components/maps/MapLayer";
+import BaseMap from "src/components/maps/components/BaseMap";
+import DrawingLayer from "src/components/maps/components/DrawingLayer";
+import MapContainer from "src/components/maps/components/MapContainer";
+import MapLayer from "src/components/maps/components/MapLayer";
 
 interface LayerMapComponentProps {
   features: FeatureCollection;
